Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,9 +2,9 @@
 #include <stdio.h>
 int main()
 {
-    int x;
+    double x;
     double area=0;
-    scanf("%d",&x);
+    scanf("%f",&x);
     printf("");
     area= (x*x)*3.14;
     printf("Area: %.2lf",area);
